+++ packages/frontend/src/providers/Explorer/ExplorerProvider.tsx
@@ -362,11 +362,16 @@ export function reducer(
         }
         case ActionType.SET_PARAMETER: {
             return produce(state, (draft) => {
+                if (!draft.unsavedChartVersion.parameters) {
+                    draft.unsavedChartVersion.parameters = {};
+                }
                 if (action.payload.value === null) {
-                    delete draft.parameters[action.payload.key];
+                    delete draft.unsavedChartVersion.parameters[
+                        action.payload.key
+                    ];
                 } else {
-                    draft.parameters = {
-                        ...draft.parameters,
+                    draft.unsavedChartVersion.parameters = {
+                        ...draft.unsavedChartVersion.parameters,
                         [action.payload.key]: action.payload.value,
                     };
                 }
@@ -374,7 +379,7 @@ export function reducer(
         }
         case ActionType.CLEAR_ALL_PARAMETERS: {
             return produce(state, (draft) => {
-                draft.parameters = {};
+                draft.unsavedChartVersion.parameters = {};
             });
         }
         case ActionType.ADD_ADDITIONAL_METRIC: {
@@ -1401,7 +1406,7 @@ const ExplorerProvider: FC<
                 ...(isEditMode ? {} : viewModeQueryArgs),
                 dateZoomGranularity,
                 invalidateCache: minimal,
-                parameters: state.parameters,
+                parameters: unsavedChartVersion.parameters || {},
             });
             dispatch({
                 type: ActionType.SET_PREVIOUSLY_FETCHED_STATE,
@@ -1418,12 +1423,12 @@ const ExplorerProvider: FC<
     }, [
         unsavedChartVersion.metricQuery,
         unsavedChartVersion.tableName,
+        unsavedChartVersion.parameters,
         projectUuid,
         isEditMode,
         viewModeQueryArgs,
         dateZoomGranularity,
         minimal,
-        state.parameters,
     ]);
 
     useEffect(() => {