+++ packages/frontend/src/providers/Explorer/defaultState.ts
@@ -43,5 +43,4 @@ export const defaultState: ExplorerReduceState = {
             isOpen: false,
         },
     },
-    parameters: {},
 };